🐛 Bug Fix: Add a Unique identifier to the GamePlayState.
The bloc does not emit a new state when the old state is the same state. When a word is found consecutively, the GamePlayState is emitted with `wordFound` the first time, and the second time around, the emitter doesn't work because the GamePlayState with `wordFound` is still the same as the previous state.