✨ Feat: Add crossword clues button and dialog
A new button, `CrosswordCluesButton`, has been added to the game UI. This button is conditionally rendered and will only appear if the current level contains crossword puzzles.

When pressed, it opens a `CrosswordCluesDialog` which displays the clues for the crosswords in the current level. The dialog gets the necessary level data from the `GamePlayBloc`.